Adapt PulseAudio to ALSA API changes.
With ALSA 1.1.9, it became fatal to include headers from /usr/include rather than /usr/include/alsa, and the headers in /usr/include added a warning regarding this. Pulseaudio also doesn't use pkg-config for a large majority of it's headers (reported upstream), so a modification had to be done to account for that as well.

git-svn-id: svn://svn.linuxfromscratch.org/BLFS/trunk@21624 af4574ff-66df-0310-9fd7-8a98e5e911e0