Update to Samba-4.10.4
This was classified as an emergency security update upstream. We recommend updating to Samba-4.10.4 or later as soon as possible, since this affects authentication response and file services from versions 4.0.2 onwards. This version should be drop-in compatible with versions 4.8.x and later.

At minimum, this will affect the 'Standard File / Print Server' configuration that we have listed in the book - but even if you only use the Client services (or GVFS), you should apply this update.

git-svn-id: svn://svn.linuxfromscratch.org/BLFS/trunk@21659 af4574ff-66df-0310-9fd7-8a98e5e911e0