Fix building kf5 with qt-5.15.1
In the "purpose" package, there is a line, which is only compiled
if Qt5 version is less or equal to 5.15.0. The reason is that it
uses an obsolete function. The problem is that the replacement
function has not been implemented yet even in 5.15.1. This generates
an error:
‘class QLocalSocket’ has no member named ‘socketError’
So change the test so that the line is compiled if version is
less than 5.16.0 (using a sed)...


git-svn-id: svn://svn.linuxfromscratch.org/BLFS/trunk@23714 af4574ff-66df-0310-9fd7-8a98e5e911e0